Rename config -> api
