+++ app/js/components/RaffleStats.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import useAsync from "react-use/lib/useAsync";
 
 const RaffleStats = () => {
   return <div>hello world!</div>;
